backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.4
asyncpg==0.30.0
alembic==1.14.1
redis[hiredis]==5.2.1
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
PyJWT==2.9.0
bcrypt==4.2.1
